refactor : CoinResponse 위치 변경
