+++ cotobang/src/main/java/com/example/cotobang/dto/CoinResponse.java
@@ -1,4 +1,4 @@
-package com.example.cotobang.domain;
+package com.example.cotobang.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
